Async and cleanup
Converted a couple more methods in the Light class to be async, although I'm not 100% sold on this method. i think perhaps I'd like to convert everything other than the initial light discovery process over to using a GLib.SocketClient so that I can take advantage of the async methods there instead.

I also noticed that the entire codebase is extremely crufty and could use a complete scrub. This is mainly due to the fact that I only ever work on it in small bursts, and tend to prioritize features over clean code...but I'll get it all sorted out before moving too far forward.