Start of big refactor
The old, tightly coupled, bad code was getting way too unwieldy. It was a good learning effort, but I need to fix the architecture and make things much cleaner before I continue to add features.

So far, I've moved the ViewModel out of it's old home, and have re-worked the left pane to use the new structure. So far, it's a bit nicer, but there is a lot more debt to pay down.